+++ packages/usc/src/index.ts
@@ -2,3 +2,13 @@
 
 export { convertTitle } from "./converter.js";
 export type { ConvertOptions, ConvertResult } from "./converter.js";
+
+export {
+  downloadTitles,
+  buildDownloadUrl,
+  buildAllTitlesUrl,
+  releasePointToPath,
+  CURRENT_RELEASE_POINT,
+  USC_TITLE_NUMBERS,
+} from "./downloader.js";
+export type { DownloadOptions, DownloadResult, DownloadedFile, DownloadError } from "./downloader.js";